% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.list.snippets_tbl}
\alias{as.list.snippets_tbl}
\title{List Conversion from Snippet Tibble}
\usage{
\method{as.list}{snippets_tbl}(x, ...)
}
\arguments{
\item{x}{"snippets_tbl" object to convert}

\item{...}{For compatibility, not in use}
}
\value{
A List
}
\description{
S3 method for convert "snippets_tbl" object to list
}
\examples{
path_vs_r <- snipper::snipper_example("vscode/r.code-snippets")

l <- as.list(read_snip_vscode(path_vs_r))
str(l)
}
