% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_snippets_tbl}
\alias{as_snippets_tbl}
\title{Convert List to Snippet Tibble}
\usage{
as_snippets_tbl(ls)
}
\arguments{
\item{ls}{A list to be convert to snippets tibble (see @details and @examples)}
}
\value{
A tibble with subclass "snippets_tbl" containing 6 columns, as per \href{https://code.visualstudio.com/docs/editor/userdefinedsnippets#_snippet-syntax}{VS code snippet syntax}
}
\description{
Convert List to Snippet Tibble
}
\details{
Names of elements in list will convert to \strong{\code{snippet_name}}, and each elements
are sub-list containing character vectors:
\itemize{
\item \strong{\code{scope}:} scope property that takes one or more \href{https://code.visualstudio.com/docs/languages/identifiers}{language identifiers}, which makes the snippet available only for those specified languages.
\item \strong{\code{prefix}:} one or more trigger words that display the snippet in IntelliSense.
\item \strong{\code{body}:} one or more lines of content, which will be joined as multiple lines upon insertion.
\item \strong{\code{description}:} an optional description of the snippet displayed by IntelliSense.
}
}
\section{Functions}{
\itemize{
\item \code{as_snippets_tbl}: list to a snippets tibble where each elements of list represent each snippets.
}}

\examples{
ls_ex <- list(
  hello = list(
   scope = "markdown",
   prefix = "hello",
   body = c("hello, there", "$0"),
   description = "Say Hello"
  ),
  world = list(
   prefix = "world",
   body = "Welcome to a new world!"
  )
)
as_snippets_tbl(ls_ex)
}
