% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_snippets_tbl}
\alias{as_snippets_tbl}
\title{Convert List or Data Frame to Snippet Tibble}
\usage{
as_snippets_tbl(x)
}
\arguments{
\item{x}{A \code{list} or \code{data.frame} to be convert to a snippets tibble (see details and examples)}
}
\value{
A tibble with subclass "snippets_tbl" containing 6 columns, based on \href{https://code.visualstudio.com/docs/editor/userdefinedsnippets#_snippet-syntax}{VS code snippet syntax}
}
\description{
Convert List or Data Frame to Snippet Tibble
}
\details{
\strong{List Input}

If the input is a list, each sub-list will map to each snippets.
Names of sub-list will convert to \strong{\code{snippet_name}}, and the named value in the sub-list
are "snippet fields" (\code{scope}, \code{prefix}, \code{body}, \code{description}) as described below.

\strong{Data Frame Input}

If the input is a data frame, \strong{row-wise mapping} to snippets will be performed.
The data frame must has columns names: \code{snippet_name}, \code{scope}, \code{prefix}, \code{body}, and \code{description} as "snippet fields".

\strong{Snippet fields}
\itemize{
\item \strong{\code{snippet_name}:} name of snippet
\item \strong{\code{scope}:} scope property that takes one or more \href{https://code.visualstudio.com/docs/languages/identifiers}{language identifiers}, which makes the snippet available only for those specified languages.
\item \strong{\code{prefix}:} one or more trigger words that display the snippet in IntelliSense.
\item \strong{\code{body}:} one or more lines of content, which will be joined as multiple lines upon insertion.
\item \strong{\code{description}:} an optional description of the snippet displayed by IntelliSense.
}
}
\section{Functions}{
\itemize{
\item \code{as_snippets_tbl}: function to convert list or data frame to a snippets tibble.
}}

\examples{
# List Input

ls_ex <- list(
  hello = list(
   scope = "markdown",
   prefix = "hello",
   body = c("hello, there", "$0"),
   description = "Say Hello"
  ),
  world = list(
   prefix = "world",
   body = "Welcome to a new world!"
  )
)

as_snippets_tbl(ls_ex)

# Data Frame Input

tbl_ex <- tibble::tibble(
  snippet_name = c(rep("hello", 2), rep("world", 2)),
  scope = "markdown",
  prefix = snippet_name,
  body = c("hello", "how are you", "Welcome !", "To a new world!"),
  description = NA
)
tbl_ex

## Body of each snippets will expand 2 lines.
as_snippets_tbl(tbl_ex)
}
