% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscode-read.R
\name{read_vscode_snippets}
\alias{read_vscode_snippets}
\title{Read Visual Studio Code Snippets files}
\usage{
read_vscode_snippets(
  paths,
  recurse = FALSE,
  regexp = "\\\\.(code-snippets)|(json)$",
  ...
)
}
\arguments{
\item{paths}{Path to VS code snippet files or directory containing those files.}

\item{recurse}{If \code{paths} contain directory, \code{recurse = TRUE} recurse into directory fully, if a positive number the number of levels to recurse.}

\item{regexp}{A regular expression passed on to \code{grep()} to filter paths. The default is to search \emph{user-defined snippets} with suffix \code{.code-snippets} (Multi-language) or \code{.json} (language-specific)}

\item{...}{passed to \code{fs::dir_ls()}}
}
\value{
A tibble with subclass "snippets_tbl" containing 6 columns, as per \href{https://code.visualstudio.com/docs/editor/userdefinedsnippets#_snippet-syntax}{VS code snippet syntax}:
\itemize{
\item \strong{\code{file_name}:} snippet file names
\item \strong{\code{snippet_name}:} names of each snippets
\item \strong{\code{scope}:} scope property that takes one or more \href{https://code.visualstudio.com/docs/languages/identifiers}{language identifiers}, which makes the snippet available only for those specified languages.
\item \strong{\code{prefix}:} one or more trigger words that display the snippet in IntelliSense.
\item \strong{\code{body}:} one or more lines of content, which will be joined as multiple lines upon insertion.
\item \strong{\code{description}:} an optional description of the snippet displayed by IntelliSense.
}
}
\description{
Read \href{https://code.visualstudio.com/docs/editor/userdefinedsnippets#_snippet-syntax}{VS code snippets} into a tibble.
}
\details{
This function read VS code snippets from a JSON format into a tibble, each rows corresponding to each snippets.
}
\examples{
path_vs_meet <- snipper::snipper_example("vscode/meet.code-snippets")
read_vscode_snippets(path_vs_meet)
}
