% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscode-write.R
\name{show_snip_vscode}
\alias{show_snip_vscode}
\title{Show Visual Studio Code Snippet}
\usage{
show_snip_vscode(snp)
}
\arguments{
\item{snp}{(snippets_tbl) A snippet tibble}
}
\value{
An "json" object
}
\description{
Show VS Code snippet as JSON.
}
\examples{
list(
 hello = list(
   scope = "markdown",
   prefix = "hello",
   body = c("hello, there", "$0"),
   description = "Say Hello"
   )
 ) \%>\%
 as_snippets_tbl() \%>\%
 show_snip_vscode()
}
