% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snipper-package.R
\docType{package}
\name{snipper-package}
\alias{snipper}
\alias{snipper-package}
\title{snipper: Snippet Converter}
\description{
Read snippets into R and convert them to various formats.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
