% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{snipper_example}
\alias{snipper_example}
\title{Get path to snipper example}
\usage{
snipper_example(path = NULL, return_path = FALSE)
}
\arguments{
\item{path}{Path to file or folder relative to \code{inst/extdata}.If \code{NULL} or \code{path} is a folder, list the file contents.
If it is a file, return absolute file path.}

\item{return_path}{If \code{TRUE}, force to return absolute path of file or folder of \code{path}.}
}
\value{
A Character vector of list files or absolute path
}
\description{
snipper comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
library(snipper)
# List Example Directory
snipper_example()
# Path of VS Code Snippet folder: `inst/extdata/vscode`
snipper_example("vscode", return_path = TRUE)
# Path of VS Code Snippet file
snipper_example("vscode/meet.code-snippets")
}
