% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscode-write.R
\name{write_snip_vscode}
\alias{write_snip_vscode}
\title{Write Visual Studio Code Snippet}
\usage{
write_snip_vscode(snp, file, ext = "code-snippets", overwrite = FALSE)
}
\arguments{
\item{snp}{(snippets_tbl) A snippet tibble to write to disk.}

\item{file}{Snippet file to write to. If provided only file name, file extension will be appended from \code{ext}.}

\item{ext}{File extension to append to \code{file}. The default is \code{.code-snippets} (Multi-language, user-defined snippets).}

\item{overwrite}{Whether to overwrite the existing file.}
}
\value{
An "json" object (invisibly)
}
\description{
Write VS Code snippet to a file.
}
\examples{
.old_wd <- setwd(tempdir())

snip_tbl <- list(
 hello = list(
   scope = "markdown",
   prefix = "hello",
   body = c("hello, there", "$0"),
   description = "Say Hello"
  )
) \%>\%
 as_snippets_tbl()

write_snip_vscode(snip_tbl, "hello")

setwd(.old_wd)

}
\seealso{
\href{https://code.visualstudio.com/docs/editor/userdefinedsnippets#_create-your-own-snippets}{VS Code Snippets}
}
